From my frustrated comment while trying to write parfib_dist:
----

And the wall is hit here: how can we make a value of type Par an instance of 'Binary'?

By default, CloudHaskell allows values of IO a, ProcessM a, and TaskM a to be 'remotable'd, so perhaps we could shoehorn in our particular monad? This privilege for specific monads is already fairly brittle, though, as TH looks for those specific types when it generates remotable code.

I'm not sure that moving to a direct scheduler where our 'Par' is an instance of MonadIO would help here, since we still will have ContT and ReaderT in the middle of our monad stack, muddying the 'Binary' instance.