Started branch for converting Direct into distributed, rather than trying to do implementations side-by-side. The main problem was that the monad type that Par should be parameterized by also ends up deep in the type of IVar.
Specifically, when IVarContents are blocked, we have a list of continuations, each of which must contain the monad where they're meant to run. This will likely take some clever engineering or an API redesign to resolve, but for now I need to stop spending time on it and just focus on the distributed part.