Stuck again. Too tired to do so now, but need to be more clever about how to type closures that are arguments to longspawn.
The essence of the problem: longSpawn needs to return an (IVar a), but where does that a come from? It needs to come from the return type of the closure being spawned.

However, when we're stealing work, we have no idea what the return type of our stolen work will be. And yet we need to be able to call the appropriate Binary instance to send the result back. 

If we put the answer-returning code on the stealer's side, then we need to send across a Closure (ProcessM a), and we run into trouble.

If we put the answer-returning code on the stealee's side, then we package things up as a Closure (ProcessM ()), and then run into trouble again.